agents/adk_orchestrator/requirements.txt
Flask>=2.0.0
gunicorn>=20.1.0
requests>=2.28.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
# Pin to the official v1.4.2 source code release from GitHub
https://github.com/google/adk-python/archive/refs/tags/v1.4.2.tar.gz
